% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FizzBuzz.R
\name{fizz_buzz}
\alias{fizz_buzz}
\title{Takes an integer (n)}
\usage{
fizz_buzz(n)
}
\arguments{
\item{a}{A real number}
}
\value{
Returns the integer, "Fizz," "Buzz," or "Fizz Buzz" based on how n is divisible
}
\description{
If n is evenly divisible by 3, the output will say "Fizz"
If n is evenly divisible by 5, the output will say "Buzz"
If n is evenly divisible by 3 and 5, the output will say "Fizz Buzz"
}
\examples{
fizz_buzz(3)
fizz_buzz(15)
fizz_buzz(25)
}
