% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flagstaff_Weather.R
\docType{data}
\name{Flagstaff_Weather}
\alias{Flagstaff_Weather}
\title{Weather in Flagstaff, AZ at the Flagstaff/ Pulliam Airport from 1950-2019.}
\format{
a data frame with 25449 observations
\describe{
  \item{Date}{The date of observation as a POSIX date format.}
  \item{TMax}{Daily maximum temperature in degrees Farhenheit.}
  \item{TMin}{Daily minimum temperature in degrees Farhenheit.}
  \item{Prcp}{Daily rainfall in inches.}
  \item{Snow}{Daily snowfall in inches.}
}
}
\source{
\url{https://www.ncdc.noaa.gov/cdo-web/search}
}
\usage{
Flagstaff_Weather
}
\description{
Weather in Flagstaff, AZ at the Flagstaff/ Pulliam Airport from 1950-2019.
}
\keyword{datasets}
